// src/types.ts
export type TypeOptions =
  | 'info'
  | 'success'
  | 'warning'
  | 'error'
  | 'default'
  | 'dark';

export type ToastPosition =
  | 'top-right'
  | 'top-center'
  | 'top-left'
  | 'bottom-right'
  | 'bottom-center'
  | 'bottom-left';

export interface ToastContentProps {
  closeToast?: () => void;
  toastProps: ToastProps;
}
export type ToastContent =
  | ReactNode
  | ((props: ToastContentProps) => ReactNode);
export interface Toast {
  content: ToastContent;
  props: ToastProps;
}

export type Id = number | string;
export type ToastTransition =
  | FC<ToastTransitionProps>
  | ComponentClass<ToastTransitionProps>;
/**
 * ClassName for the elements - can take a function to build a classname or a raw string that is cx'ed to defaults
 */
export type ClassName =
  | ((context?: {
      defaultClassName?: string;
      position?: ToastPosition;
      rtl?: boolean;
    }) => string)
  | string;
export type ToastClassName =
  | ((context?: {
      type?: TypeOptions;
      defaultClassName?: string;
      position?: ToastPosition;
      rtl?: boolean;
    }) => string)
  | string;

export interface ClearWaitingQueueParams {
  containerId?: Id;
}

export type DraggableDirection = 'x' | 'y';

interface CommonOptions {
  /**
   * Pause the timer when the mouse hover the toast.
   * `Default: true`
   */
  pauseOnHover?: boolean;

  /**
   * Pause the toast when the window loose focus.
   * `Default: true`
   */
  pauseOnFocusLoss?: boolean;

  /**
   * Remove the toast when clicked.
   * `Default: true`
   */
  closeOnClick?: boolean;

  /**
   * Set the delay in ms to close the toast automatically.
   * Use `false` to prevent the toast from closing.
   * `Default: 5000`
   */
  autoClose?: number | false;

  /**
   * Set the default position to use.
   * `One of: 'top-right', 'top-center', 'top-left', 'bottom-right', 'bottom-center', 'bottom-left'`
   * `Default: 'top-right'`
   */
  position?: ToastPosition;

  /**
   * Pass a custom close button.
   * To remove the close button pass `false`
   */
  closeButton?: ReactElement | ((props: any) => ReactElement) | boolean;

  /**
   * An optional css class to set for the progress bar.
   */
  progressClassName?: ToastClassName;

  /**
   * An optional style to set for the progress bar.
   */
  progressStyle?: StyleProp<ViewStyle>;

  /**
   * An optional css class to set for the toast content.
   */
  bodyClassName?: ToastClassName;

  /**
   * An optional inline style to apply for the toast content.
   */
  bodyStyle?: StyleProp<ViewStyle>;

  /**
   * Hide or show the progress bar.
   * `Default: false`
   */
  hideProgressBar?: boolean;

  /**
   * Pass a custom transition built with react-transition-group.
   */
  transition?: ToastTransition;

  /**
   * Allow toast to be draggable
   * `Default: true`
   */
  draggable?: boolean;

  /**
   * The percentage of the toast's width it takes for a drag to dismiss a toast
   * `Default: 80`
   */
  draggablePercent?: number;

  /**
   * Specify in which direction should you swipe to dismiss the toast
   * `Default: "x"`
   */

  draggableDirection?: DraggableDirection;

  /**
   * Define the ARIA role for the toast
   * `Default: alert`
   *  https://www.w3.org/WAI/PF/aria/roles
   */
  role?: string;

  /**
   * Set id to handle multiple container
   */
  containerId?: Id;

  /**
   * Fired when clicking inside toaster
   */
  onClick?: (event: MouseEvent) => void;
  /**
   * Fired when clicking inside toaster
   */
  onPress?: () => void;

  /**
   * Support right to left display.
   * `Default: false`
   */
  rtl?: boolean;
}

export interface ToastOptions extends CommonOptions {
  /**
   * An optional css class to set.
   */
  className?: ToastClassName;

  /**
   * Called when toast is mounted.
   */
  onOpen?: <T = {}>(props: T) => void;

  /**
   * Called when toast is unmounted.
   */
  onClose?: <T = {}>(props: T) => void;

  /**
   * An optional inline style to apply.
   */
  style?: StyleProp<ViewStyle>;

  /**
   * An optional style to apply to text.
   */
  textStyle?: StyleProp<TextStyle>;

  /**
   * Set the toast type.
   * `One of: 'info', 'success', 'warning', 'error', 'default'`
   */
  type?: TypeOptions;

  /**
   * Set a custom `toastId`
   */
  toastId?: Id;

  /**
   * Used during update
   */
  updateId?: Id;

  /**
   * Set the percentage for the controlled progress bar. `Value must be between 0 and 1.`
   */
  progress?: number | string;

  /**
   * Add a delay in ms before the toast appear.
   */
  delay?: number;
}